// src/shared/types/maps.ts
/* eslint-disable @typescript-eslint/naming-convention */

/* Map code (cs). */
export type MapCode =
  | 'de_ancient'
  | 'de_anubis'
  | 'de_cbble'
  | 'de_dust2'
  | 'de_inferno'
  | 'de_mirage'
  | 'de_nuke'
  | 'de_overpass'
  | 'de_train'
  | 'de_vertigo';

/* Map name (cs). */
export type MapName =
  | 'Ancient'
  | 'Anubis'
  | 'Cbble'
  | 'Dust2'
  | 'Inferno'
  | 'Mirage'
  | 'Nuke'
  | 'Overpass'
  | 'Train'
  | 'Vertigo';